Improve update checker
